pyproject.toml
[project]
dynamic = ["version"]
name = "vision.cpp"
requires-python = ">=3.12"
dependencies = ["pillow"]

[dependency-groups]
dev = [
    "torch",
    "torchvision",
    "timm",
    "pytest",
    "numpy",
    "opencv-python",
    "ruff",
    "einops>=0.8.1",
    "spandrel>=0.4.1",
    "gguf>=0.17.1",
    "boto3~=1.39.0",
]

[tool.uv]
package = false
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[tool.ruff]
target-version = "py312"
line-length = 100
preview = true

[tool.ruff.lint]
ignore = ["E741"]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
